MDEV-27677: Implement JSON_OVERLAPS()
If both the arguments are json objects or json arrays, then compare each
key or array element with other key or array element until match is
found. If both the json documents are of object type, return true if
at least one matching key-value pair is found. If both json documents are
of array type, return true if at least one array element matches completely
with other. If one of the json document is object and other is array or
vice versa, return true if object completely matches with array element
of object type. If match is found then return true else return false.
For arguements not of object or array type, return true if the value and
value type is same.