MDEV-27740: Incorrect SST with non-standard directories for innodb files
Prior to the MDEV-27524 change, a joiner node could not join the cluster
if it stored redo logs in a directory other than the default data directory.
The final version of MDEV-27524 should fix this shortcoming, but we need
new tests to verify correct operation in such scenarios.

1) This commit adds tests to verify that explicitly specifying
all directories for innodb works fine when using Galera with
SST via mariabackup and rsync, i.e. to verify that the previous
changes from the final version of the MDEV-27524 fixed all bugs
that caused incorrect operation of the SST scripts and the
mariabackup utility with innodb directories.

2) To check all this, a new control command has been added
to the mtr, which is specified as a pragma in any of the sections
(groups) in the .cnf configuration files:

   #!bootstrap-opts <option1>[, <option2>]...

This feature allows you to list the names of the parameters
located in this section of the .cnf file, which must be converted
into command line options and passed to the input of the bootstrap
process during the construction of the initial base for the next
test run.

This allows you to specify the parameters that should be passed
to the bootstrap process directly in the .cnf files, without creating
separate .opt files (this is especially useful for the multi-node
configurations).