MDEV-27760 event may non stop replicate in circular semisync setup
MDEV-21117 had to relax own events acceptance condition for a case
when a former semisync master server recovers after crash as the
semisync slave.  That however admitted a possibility for endless event
"orbiting" in the non-strict slave gtid mode of semisync circular
setup.

The same server-id event termination is restored now for the
non-strict gtid mode to follow regular rules (that is it's ignored
unless @@global.replicate_same_server_id allows it in).

In the strict gtid mode, to address MDEV-21117 recovery concern, a
transaction gets accepted if its gtid is ordered strictly greater than
the current slave gtid state.
When gtid is a part of the state (ordered less or equal the latest
gtid of the state) the same-server-id ignore rules apply as usual.
The gtid strict mode is safe to accept transactions even if
the slave state were not set correct by the user.
An added test shows a typical out-of-order error at execution so
no data corruption is guaranteed.